chore: move enum package
