+++ src/main/java/com/example/cheerdo/todo/service/TodoServiceImpl.java
@@ -7,7 +7,7 @@ import com.example.cheerdo.todo.dto.request.GetTodoRequestDto;
 import com.example.cheerdo.todo.dto.request.ModifyTodoRequestDto;
 import com.example.cheerdo.todo.dto.request.WriteTodoRequestDto;
 import com.example.cheerdo.todo.dto.response.TodoResponseDto;
-import com.example.cheerdo.todo.enums.Type;
+import com.example.cheerdo.entity.enums.Type;
 import com.example.cheerdo.repository.TodoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
